requirements.txt
django==1.7.7

pyyaml==3.11 # Used for fixtures
django_compressor # This allows us to use Coffeescript and SCSS
django-libsass==0.2 # Plugin for compressor to use SCSS
djangorestframework==3.1.1 # For passing models to javascript
django-sekizai==0.8.1 # For adding to parent blocks


# Authentication

django-allauth==0.18.0 # Authentication and social auth